Nke oom fix

See merge request iot/cumulocity/common/wip/cumulocity-lora!14

+++ lora-codec-nke/src/main/java/lora/codec/nke/C8YDataType.java
@@ -7,6 +7,6 @@ import org.joda.time.DateTime;
 import lora.codec.C8YData;
 
 public interface C8YDataType {
-    public void sendToCumulocity(ManagedObjectRepresentation device, C8YData c8yData, int value, DateTime time);
+    public void sendToCumulocity(ManagedObjectRepresentation device, C8YData c8yData, float value, DateTime time);
     public C8YDataType setAsDeviceProperty(String propertyName);
 }
